Incorporate more code review feedback - Separated the html and kml generation into different scripts - Merged the html header and footer into a single file - Used a string template to insert the generated housing table into the html file - Added wait time to prevent exceeding the google map query limit - Updated README and sample outputs
