--- affordable-housing/housing_script_footer.txt
@@ -1,3 +1,0 @@
-</table>
-</body>
-</html>
